fix: template generation now works with llm_client
fix: chat handler bug fixes
