+++ src/utils/chat/chatHandlers.js
@@ -125,7 +125,6 @@ export const handleChat = async (
                         content: streamedContent, // Update content cumulatively
                         context: streamedContext, // Update context if received
                         loading: !finalResponseReceived, // Set loading to false only when 'end' is received
-                        isThinkingExpanded: false, // Always start with thinking collapsed
                     };
                     // If stream finished, remove loading and potentially set final state
                     if (finalResponseReceived) {
